Изменение : Изменен механизм расчета/уменьшения Dim(применяется для расчета соотношения сторон). За основу взят код из ffmpeg, т.к. он более "интелектуальный". Этим самым избавляемся от "диких" цифр в AR;
git-svn-id: svn://mpc-next.ru/mpc-be/trunk@4834 d191f3aa-8897-45be-b9a4-56e5fb472b67

+++ src/filters/parser/RawVideoSplitter/RawVideoSplitter.cpp
@@ -196,8 +196,8 @@ HRESULT CRawVideoSplitterFilter::CreateOutputs(IAsyncReader* pAsyncReader)
 		int    height		= 0;
 		int    fpsnum		= 24;
 		int    fpsden		= 1;
-		int    sar_x		= 1;
-		int    sar_y		= 1;
+		LONG   sar_x		= 1;
+		LONG   sar_y		= 1;
 		FOURCC fourcc		= FCC('I420'); // 4:2:0 - I420 by default
 		FOURCC fourcc_2		= 0;
 		WORD   bpp			= 12;
